Organized code in the analyze folder
This is work towards #803

We move some functions into some logical locations in the
codebase:

- Move analysis of container images into default/container.
- Move analysis of Dockerfiles into default/dockerfile.
- Move some functions into their respective places in the files.
  There are definitely more functions to move around, but at
  this point, the codebase is so broken this commit should be
  considered dangerous to apply without the whole patchset.
