+++ tern/analyze/default/container/collect.py
