+++ tern/analyze/default/container/filter.py
