+++ tern/analyze/default/container/image.py
@@ -4,20 +4,27 @@
 # SPDX-License-Identifier: BSD-2-Clause
 
 """
-Analyze a Docker container image
+Analyze the container image in default mode
 """
 
+import docker
 import logging
+import subprocess  # nosec
 import sys
 
+from tern.classes.notice import Notice
+from tern.classes.docker_image import DockerImage
 from tern.report import errors
 from tern.utils import constants
 from tern.utils import rootfs
-from tern.classes.notice import Notice
 from tern.analyze import common
 import tern.analyze.docker.helpers as dhelper
 from tern.command_lib import command_lib
 from tern.analyze.docker import dockerfile as d_file
+from tern.classes.notice import Notice
+import tern.analyze.docker.helpers as dhelper
+from tern.load import docker_api
+from tern.report import formats
 
 
 # global logger
@@ -68,3 +75,42 @@ def image_setup(image_obj):
         if layer.import_str:
             layer.origins.add_notice_to_origins(origin_str, Notice(
                 'Imported in Dockerfile using: ' + layer.import_str, 'info'))
+
+
+def load_base_image():
+    '''Create base image from dockerfile instructions and return the image'''
+    base_image, dockerfile_lines = dhelper.get_dockerfile_base()
+    # try to get image metadata
+    if docker_api.dump_docker_image(base_image.repotag):
+        # now see if we can load the image
+        try:
+            base_image.load_image()
+        except (NameError,
+                subprocess.CalledProcessError,
+                IOError,
+                docker.errors.APIError,
+                ValueError,
+                EOFError) as error:
+            logger.warning('Error in loading base image: %s', str(error))
+            base_image.origins.add_notice_to_origins(
+                dockerfile_lines, Notice(str(error), 'error'))
+    return base_image
+
+
+def load_full_image(image_tag_string, digest_string):
+    '''Create image object from image name and tag and return the object'''
+    test_image = DockerImage(image_tag_string, digest_string)
+    failure_origin = formats.image_load_failure.format(
+        testimage=test_image.repotag)
+    try:
+        test_image.load_image()
+    except (NameError,
+            subprocess.CalledProcessError,
+            IOError,
+            docker.errors.APIError,
+            ValueError,
+            EOFError) as error:
+        logger.warning('Error in loading image: %s', str(error))
+        test_image.origins.add_notice_to_origins(
+            failure_origin, Notice(str(error), 'error'))
+    return test_image
