+++ tern/analyze/default/dockerfile/dockerfile.py
