+++ tern/analyze/default/dockerfile/helpers.py
